Added review changes

+++ src/main/java/com/customer/transacation/dto/TransactionDetails.java
@@ -1,5 +1,6 @@
 package com.customer.transacation.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import lombok.Data;
@@ -8,6 +9,6 @@ import lombok.Data;
 public class TransactionDetails {
     private Long transactionId;
     private LocalDate transactionDate;
-    private double amountSpent;
+    private BigDecimal amountSpent;
     private int rewardPoints;
 }
